Filter out expected mismatches for flats

+++ nirspec_pipe_testing_tool/calwebb_spec2_pytests/auxiliary_code/flattest_fs.py
@@ -5,9 +5,6 @@ import sys
 import shutil
 import numpy as np
 from astropy.io import fits
-from glob import glob
-from copy import deepcopy
-from gwcs import wcstools
 from jwst import datamodels
 from . import auxiliary_functions as auxfunc
 
@@ -535,64 +532,61 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
                                 print()
                                 input()
 
-                        # Remove all pixels with values=1 (outside slit boundaries) for statistics
-                        if pipeflat[k, j] == 1:
+                        # Remove all pixels with values=1 or 0 (outside slit boundaries)
+                        # or non-zero DQ for statistics
+                        if (pipeflat[k, j] == 1) or (pipeflat[k, j] == 0) or (pipeflat_dq[k, j] != 0):
                             delf[k, j] = np.nan
                             delflaterr[k, j] = np.nan
                             flatcor[k, j] = np.nan
                             flat_err[k, j] = np.nan
 
-            # only keep points in slit
-            delfg = delf[np.isfinite(delf)]
-            delflaterr = delflaterr[np.isfinite(delf)]
-
             # attempt remove outliers, for better statistics, only use points where pipe-calc <= 1.0
-            outliers_idx = np.where(np.absolute(delfg) <= 1.0)
+            outliers = (np.absolute(delf / flatcor) > 1.0)
 
             # if the remaining points are more than half the original number, remove outliers
-            if len(outliers_idx) >= len(delfg)/2.0:
-                delfg = delfg[outliers_idx]
-                # do the same with the error differences array
-                outliers_idx = np.where((np.absolute(delflaterr) <= 1.0))
-                delflaterr = delflaterr[outliers_idx]
-            if debug:
-                print('delf = ', np.shape(delf),  delf)
+            if np.sum(outliers) <= delf.size / 2.0:
+                delf[outliers] = np.nan
+                flatcor[outliers] = np.nan
 
-            if delf.size == 0:
+                # do the same with the error differences array
+                outliers = (np.absolute(delflaterr / flat_err) > 1.0)
+                delflaterr[outliers] = np.nan
+                flat_err[outliers] = np.nan
+
+
+            if np.all(np.isnan(delf)):
                 msg1 = " * Unable to calculate statistics because difference array has all values as NaN. Test will be set to FAILED."
                 print(msg1)
                 log_msgs.append(msg1)
                 test_result = "FAILED"
-                delfg_mean, delfg_median, delfg_std = np.nan, np.nan, np.nan
-                stats = [delfg_mean, delfg_median, delfg_std]
+                delf_mean, delf_median, delf_std = np.nan, np.nan, np.nan
+                stats = [delf_mean, delf_median, delf_std]
             else:
                 msg = "Calculating statistics... "
                 print(msg)
                 log_msgs.append(msg)
-                if delfg.size == 0:
-                    msg1 = " * Unable to calculate statistics because difference array has all outlier values. Test will be set to FAILED."
-                    print(msg1)
-                    log_msgs.append(msg1)
-                    test_result = "FAILED"
-                    delfg_mean, delfg_median, delfg_std = np.nan, np.nan, np.nan
-                    stats = [delfg_mean, delfg_median, delfg_std]
-                else:
-                    stats_and_strings = auxfunc.print_stats(delfg, "Flat Difference", float(threshold_diff), absolute=True)
-                    stats, stats_print_strings = stats_and_strings
-                    delfg_mean, delfg_median, delfg_std = stats
-                    for msg in stats_print_strings:
-                        log_msgs.append(msg)
-                    err_stats_and_strings = auxfunc.print_stats(delflaterr, "Flat Error Difference", float(threshold_diff), absolute=True)
-                    err_stats, err_stats_print_strings = err_stats_and_strings
 
-                    # This is the key argument for the assert pytest function
-                    median_diff = False
-                    if abs(delfg_median) <= float(threshold_diff):
-                        median_diff = True
-                    if median_diff:
-                        test_result = "PASSED"
-                    else:
-                        test_result = "FAILED"
+                stats_and_strings = auxfunc.print_stats(
+                    delf/flatcor, "Flat Difference",
+                    float(threshold_diff), absolute=False)
+                stats, stats_print_strings = stats_and_strings
+                delf_mean, delf_median, delf_std = stats
+                for msg in stats_print_strings:
+                    log_msgs.append(msg)
+
+                err_stats_and_strings = auxfunc.print_stats(
+                    delflaterr/flat_err, "Flat Error Difference",
+                    float(threshold_diff), absolute=False)
+                err_stats, err_stats_print_strings = err_stats_and_strings
+
+                # This is the key argument for the assert pytest function
+                median_diff = False
+                if abs(delf_median) <= float(threshold_diff):
+                    median_diff = True
+                if median_diff:
+                    test_result = "PASSED"
+                else:
+                    test_result = "FAILED"
 
             msg = " *** Result of the test: " + test_result
             print(msg)
@@ -600,11 +594,6 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
             total_test_result.append(test_result)
 
             # make histogram
-            flatcor_copy = deepcopy(flatcor)
-            flatcor_copy[~np.isfinite(flatcor)] = 1.0
-            flat_err_copy = deepcopy(flat_err)
-            flat_err_copy[~np.isfinite(flatcor)] = np.nan
-
             if show_figs or save_figs:
                 # set plot variables
                 main_title = filt + "   " + grat + "   SLIT=" + slit_id + "\n"
@@ -616,8 +605,8 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
                 info_img = [title, "x (pixels)", "y (pixels)"]
                 xlabel, ylabel = "flat$_{pipe}$ - flat$_{calc}$", "N"
                 info_hist = [xlabel, ylabel, bins, stats]
-                if delfg.size != 0 and delfg[1] is np.nan:
-                    msg = "Unable to create plot of pipeline - calulated flat values."
+                if delf.size != 0 and delf[1] is np.nan:
+                    msg = "Unable to create plot of pipeline - calculated flat values."
                     print(msg)
                     log_msgs.append(msg)
                 else:
@@ -628,17 +617,17 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
                         plt_name = os.path.join(os.getcwd(), "FS_" + det + "_" + slit_id + "_flatdiff_histogram.png")
                         print("No output_directory was provided. Figures will be saved in current working directory:")
                         print(plt_name + "\n")
-                    difference_img = (pipeflat - flatcor_copy)  # /flatcor
+                    difference_img = (pipeflat - flatcor)  # /flatcor
                     if debug:
                         print('np.shape(pipeflat), pipeflat: ', np.shape(pipeflat), pipeflat[np.where(pipeflat != 1.0)])
                         print('np.shape(flatcor), flatcor: ', np.shape(flatcor), flatcor[np.isfinite(flatcor)])
                         print('np.shape(difference_img), difference_img: ', np.shape(difference_img), ~np.isnan(difference_img))
                     # set the range of values to be shown in the image, will affect color scale
-                    if delfg_std <= 0.0:
-                        vminmax = [-5 * delfg_std, 5 * delfg_std]
+                    if delf_std <= 0.0:
+                        vminmax = [-5 * delf_std, 5 * delf_std]
                     else:
                         vminmax = None
-                    auxfunc.plt_two_2Dimgandhist(difference_img, delfg, info_img, info_hist, plt_name=plt_name,
+                    auxfunc.plt_two_2Dimgandhist(difference_img, delf, info_img, info_hist, plt_name=plt_name,
                                                  vminmax=vminmax, plt_origin=plt_origin,
                                                  show_figs=show_figs, save_figs=save_figs)
 
@@ -660,7 +649,7 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
                         plt_name = os.path.join(os.getcwd(), "FS_" + det + "_" + slit_id + "_flatdiff_err_histogram.png")
                         print("No output_directory was provided. Figures will be saved in current working directory:")
                         print(plt_name + "\n")
-                    difference_img = pipeflat_err - flat_err_copy
+                    difference_img = pipeflat_err - flat_err
                     auxfunc.plt_two_2Dimgandhist(difference_img, delflaterr, info_img, info_hist, plt_name=plt_name,
                                                  vminmax=None, plt_origin=plt_origin,
                                                  show_figs=show_figs, save_figs=save_figs)
@@ -681,11 +670,10 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
                 log_msgs.append(msg)
 
                 # this is the file to hold the image of calculated flat correctoion values
-                #flatcor_copy[np.isnan(flatcor_copy)] = 1.0
-                outfile_ext = fits.ImageHDU(flatcor_copy, name=slit_id+'_SCI')
+                outfile_ext = fits.ImageHDU(flatcor, name=slit_id+'_SCI')
                 outfile.append(outfile_ext)
                 # this is the file to hold the corresponding error extensions
-                outfile_ext = fits.ImageHDU(flat_err_copy, name=slit_id+'_ERR')
+                outfile_ext = fits.ImageHDU(flat_err, name=slit_id+'_ERR')
                 outfile.append(outfile_ext)
 
                 # this is the file to hold the image of pipeline-calculated difference values
@@ -703,8 +691,10 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
             auxfunc.calc_flat_total_slt_err(flat_field_pipe_outfile, slit_id,
                                             flout_slt_sci, flout_slt_err,
                                             input_sci, input_err, input_var_psn, input_var_rnse,
-                                            pipeflat, pipeflat_err, flatcor_copy, flat_err_copy,
+                                            pipeflat.copy(), pipeflat_err.copy(),
+                                            flatcor.copy(), flat_err.copy(),
                                             show_plts=show_figs, save_plts=save_figs)
+        break
 
     # close datamodels
     flatfile.close()
@@ -712,8 +702,8 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
     model.close()
 
     if writefile:
-        outfile_name = flat_field_pipe_outfile.replace("interpolatedflat.fits", "flat_calc.fits")
-        complfile_name = flat_field_pipe_outfile.replace("interpolatedflat.fits", "flat_comp.fits")
+        outfile_name = flat_field_pipe_outfile.replace("flat_field", "flat_calc")
+        complfile_name = flat_field_pipe_outfile.replace("flat_field", "flat_comp")
 
         # create the fits list to hold the calculated flat values for each slit
         outfile.writeto(outfile_name, overwrite=True)
@@ -721,6 +711,7 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
         # this is the file to hold the image of pipeline-calculated difference values
         complfile.writeto(complfile_name, overwrite=True)
 
+        print('')
         msg = "Fits file with calculated flat values of each slit saved as: "
         print(msg)
         log_msgs.append(msg)
@@ -754,9 +745,12 @@ def flattest(step_input_filename, dflat_path, sflat_path, fflat_path, writefile=
         result_msg = "One or more slits FAILED flat_field test."
 
     # remove the copied files
-    os.remove(pipe_interpolated_flat_file)
-    os.remove(flat_field_pipe_outfile)
-    os.remove(wcs_file)
+    if os.path.isfile(pipe_interpolated_flat_file):
+        os.remove(pipe_interpolated_flat_file)
+    if os.path.isfile(flat_field_pipe_outfile):
+        os.remove(flat_field_pipe_outfile)
+    if os.path.isfile(wcs_file):
+        os.remove(wcs_file)
 
     # end the timer
     flattest_end_time = time.time() - flattest_start_time
